refactor(matrix-viewer): extract helpers + atoms
